[scanner] fix: split oversized drilldown and events views
Continues the drilldown-view split with the remaining oversized files:

- DeploymentDrillDown: extract all kubectl fetch/scale/refresh logic
  into useDeploymentData() (195 lines, down from 437)
- Events.tsx: extract EventsOverviewTab, EventsTimelineTab, EventsListTab,
  and useEventsData() into events/events-page/ (235 lines, down from 670)
- EventsPanel.tsx: extract EventsPanelHeader, EventsPanelList,
  BatchProcessingBanner, Group/EmptyState primitives, and
  useEventsPanelData() into stellar/events-panel/ (154 lines, down from 500)

No behavior change; all public exports and rendered output are preserved.

Fixes #21506
Fixes #21620
Fixes #21621
Fixes #21623
Fixes #21624
