+++ web/src/components/drilldown/views/deployment-drilldown/index.ts
@@ -13,3 +13,4 @@ export type { DeploymentHeaderProps } from './DeploymentHeader'
 export { DeploymentTabs } from './DeploymentTabs'
 export type { DeploymentTabsProps, DeploymentTabItem } from './DeploymentTabs'
 export { useCopyFeedback } from './useCopyFeedback'
+export { useDeploymentData } from './useDeploymentData'
